+++ src/main.cpp
@@ -6,6 +6,7 @@
 #include "managers/displaymanager.h"
 #include "managers/networkmanager.h"
 #include "managers/rtcmanager.h"
+#include "managers/screenmanager.h"
 #include "managers/sensormanager.h"
 #include "managers/stylemanager.h"
 #include "managers/taskmanager.h"
@@ -13,7 +14,6 @@
 #include "screens/lockscreen.h"
 #include "screens/mainscreen.h"
 #include "screens/samplingsettingsscreen.h"
-#include "screens/screenmanager.h"
 #include "screens/settingsscreen.h"
 #include "screens/timesettingsscreen.h"
 #include "screens/wifiscreen.h"
